chore(app): implement client-side filters storage persistency
Updated the Monitoring component and useMonitoringSession composable to support persistent storage of filters using the browser's local storage. This change ensures that user preferences for event types and names are retained across sessions, improving the usability of the monitoring interface. Filters are now initialized from local storage and updated reactively on changes.