chore(app): update Nuxt configuration and dependency management
Updated the Nuxt configuration by adjusting the modules and storybook settings, and refined dependency management in the `vite` configuration. Disabled the `sonarjs/no-clear-text-protocols` rule to accommodate specific requirements, commented out the storybook module to disable it temporarily, and expanded the `exclude` and `include` lists in the `vite` config to ensure better handling of dependencies during the build process. These changes streamline configuration and improve build performance.