chore(dependencies): update package dependencies across multiple modules
Updated various package dependencies to their latest versions in the `app`, `core`, `module-core`, `module-google`, and `module-piste-gouv-fr` modules. This includes minor and patch version updates for better compatibility and to incorporate the latest bug fixes and performance improvements. This will help ensure that the project stays up-to-date with the current libraries, reducing the likelihood of facing issues related to outdated dependencies.